# web/openapi.yaml
openapi: 3.0.0
info:
  title: 'ZSF API'
  version: '1.1'
servers:
  -
    url: /api/1.1
    description: 'ZSF API'
paths:
...
  /Roles/Add:
    post:
      tags:
        - Roles
      summary: 'Add new role to system'
      description: 'Add new role to system'
      operationId: addRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { created: { type: string }, id: { type: number }, name: { type: string } }
                  type: object
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  /Roles:
    get:
      tags:
        - Roles
      summary: 'Retrieves all roles'
      description: 'Retrieves all roles'
      operationId: getRoles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { created: { type: string }, id: { type: number }, name: { type: string } }
                  type: object
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  '/Roles/UserRoles/{userId}':
    get:
      tags:
        - Roles
      summary: 'Retrieve user roles'
      description: 'Retrieve roles for the given user'
      operationId: getUserRoles
      parameters:
        -
          name: userId
          in: path
          description: 'unique user identifier'
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { created: { type: string }, id: { type: number }, name: { type: string } }
                  type: object
      security:
        -
          header_token: []
        -
          cookie_token: []
  /Roles/UserRoles:
    get:
      tags:
        - Roles
      summary: 'Retrieve my user roles'
      description: 'Retrieve my user roles'
      operationId: getMyRoles
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { created: { type: string }, id: { type: number }, name: { type: string } }
                  type: object
      security:
        -
          header_token: []
        -
          cookie_token: []
  /Roles/Remove:
    post:
      tags:
        - Roles
      summary: 'Remove role from system'
      description: 'Remove role from system'
      operationId: removeRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  created: { type: string }
                  id: { type: number }
                  name: { type: string }
                type: object
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  /Roles/RemoveUserRole:
    post:
      tags:
        - Roles
      summary: 'Removes role for given user'
      description: 'Removes role for given user'
      operationId: removeUserRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: number
                name:
                  type: string
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  /Roles/RemoveUserRoles:
    post:
      tags:
        - Roles
      summary: 'Removes all roles for user'
      description: 'Removes all roles for user'
      operationId: removeUserRoles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: number
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  /Roles/SetUserRole:
    post:
      tags:
        - Roles
      summary: 'Assigns role to user'
      description: 'Assigns role to user'
      operationId: setUserRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: number
                name:
                  type: string
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  created: { type: string }
                  id: { type: number }
                  name: { type: string }
                type: object
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
  /Roles/UserInRole:
    post:
      tags:
        - Roles
      summary: 'Check if user is assigned a role'
      description: 'Check if user is assigned a role'
      operationId: userInRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                userId:
                  type: number
                name:
                  type: string
              type: object
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        -
          header_token: []
        -
          cookie_token: []
  '/Roles/UsersInRole/{RoleName}':
    get:
      tags:
        - Roles
      summary: 'Retrieve any users assigned to a role'
      description: 'Retrieve any users assigned to a role'
      operationId: usersInRole
      parameters:
        -
          name: RoleName
          in: path
          description: 'Role name to search with'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { email: { description: 'User''s email address', type: string }, emailConfirmed: { description: 'If user''s email address has been confirmed', type: boolean }, id: { description: 'User''s unique identifier', type: number }, joined: { description: 'Date and time the user joined', type: string }, lastActive: { description: 'Date and time the user was last active, if available', type: string }, lastLogin: { description: 'Date and time the user last logged in, if available', type: string } }
                  type: object
      security:
        -
          admin_header_token: []
        -
          admin_cookie_token: []
...
components:
...
  securitySchemes:
    header_token:
      type: http
      scheme: bearer
    cookie_token:
      type: apiKey
      name: zsf_token
      in: cookie
    admin_header_token:
      type: http
      scheme: bearer
    admin_cookie_token:
      type: apiKey
      name: zsf_token
      in: cookie
...